Optimize Seer a bit per testing with @Meowchestra
polling rates are good at 1ms, except chat log, which was eating a ton of cpu time. Need to figure out what is wrong with that. Also, we can use a socketfilter with machina, because we don't care about losing a few packets during initial game->server connection, which drops cpu usage.

there is still a memory leak in machina that has existed since the beginning of time that may need to be fixed (it exists in ACT Parser too.). it will occur much less often with these settings though.

Co-Authored-By: Meowchestra <42126400+Meowchestra@users.noreply.github.com>