Playlist Header Improvements
* Added a return icon when a playlist is loaded so people can figure out you can click on the header to return to the playlist selection.

* Changed order to display GetTotalTime before the Playlist.GetName. Because the font is not monospaced and a playlist can have a title up to 40 characters, padding the name 75 spaces before displaying the playlist time does _not_ consistently display the time in the same right side spot.

For example if a playlist is named '1', 'ffffffffffffffffffffffffffffffffffffffff' or 'LLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLL', 'TTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTT' all 4 will have different lengths that cause the time to be in different locations. Plus padding it a static amount looks bad if the player is ever resized larger or smaller. I tried a lot of different things such as always having a set amount of whitespace characters and subtracting based on the name length to ensure the same amount of character spacing, setting a monospace font which reduces legibility and shouldn't really be done globally, and a bunch of xaml / backed code tests to set containers / columns and data binds. Nothing I tried could right align the time so I gave up and decided to just display the time before the name.