Preview Adjustments
* Merge Play/Pause button into one.

* Fixed crash when clicking load while highlighting a playlist.

If you are inside a playlist and select an item then click load in the previewer, it should load the song into the previewer. This now fixes an issue where if you highlight a playlist and click load, it won't crash bmp due to not being able to find a song.