UI Fixes & Improvements (#282)
* Fix binding errors

* SongBrowser complains because Title binding doesn't even exist. Doesn't seem used or needed so removed.

* Settings Midi device box complains with 'None' due to content vertical and horizontal aligns. Removing this still shows 'None' in the content box and no longer complains so assume not needed. Untested if midi device names show up as I do not have one.

* Playlist Header Improvements

* Added a return icon when a playlist is loaded so people can figure out you can click on the header to return to the playlist selection.

* Changed order to display GetTotalTime before the Playlist.GetName. Because the font is not monospaced and a playlist can have a title up to 40 characters, padding the name 75 spaces before displaying the playlist time does _not_ consistently display the time in the same right side spot.

For example if a playlist is named '1', 'ffffffffffffffffffffffffffffffffffffffff' or 'LLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLLL', 'TTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTTT' all 4 will have different lengths that cause the time to be in different locations. Plus padding it a static amount looks bad if the player is ever resized larger or smaller. I tried a lot of different things such as always having a set amount of whitespace characters and subtracting based on the name length to ensure the same amount of character spacing, setting a monospace font which reduces legibility and shouldn't really be done globally, and a bunch of xaml / backed code tests to set containers / columns and data binds. Nothing I tried could right align the time so I gave up and decided to just display the time before the name.

* Preview Adjustments

* Merge Play/Pause button into one.

* Fixed crash when clicking load while highlighting a playlist.

If you are inside a playlist and select an item then click load in the previewer, it should load the song into the previewer. This now fixes an issue where if you highlight a playlist and click load, it won't crash bmp due to not being able to find a song.