Program:ElectricGuitar added to instrument list
* Added Program:ElectricGuitar to instrument list so midis that have the track name Program:ElectricGuitar display that instead of ElectricGuitarOverdriven. This is so you can more easily identify which tracks/midis are using Program to tell what may or may not work with the "Ignore Program Change Events" setting, since that avoids tracks with the name Program:ElectricGuitar.

Behavior should be unaffected and function like ElectricGuitarOverdriven instrument. This change is merely for the displayed performer instrument name since the main instrument name is used in the UI, not the alternative names or track names.

VSTLoader was also updated to ignore this new instrument to avoid it breaking and trying to download a vst for something that doesn't exist. Previewing with prog changes still seem to work with the new instrument / name.