Fix seer memory leak
* Fixed memory leak in seer.

It seemed to keep adding new processes to the processes list on every iteration without clearing the list. Now the processes list is created new on every iteration like before, except with the newer sorting by pid start time.

The older version of ProcessWatcher had a potential performance issue as it added a delay of 1 millisecond for each iteration, which could result in increased CPU usage. This was changed to 100ms as a potential improvement. However the ProcessWatcher constantly iterating is not ideal as cpu cycles are constantly spiking up and down, as well as memory usage being garbage collected repeatedly causing spikes up and down.

Instead of constantly iterating to check for new processes, it may be better to use WMI API to monitor the creation of new processes. I'm just not sure on the viability of this for cross-platform WINE support.